+++ libnavui-maneuver/src/main/java/com/mapbox/navigation/ui/maneuver/view/MapboxSubManeuver.kt
@@ -61,13 +61,13 @@ class MapboxSubManeuver : AppCompatTextView {
      * @param maneuver SubManeuver
      */
     @Deprecated(
-        message = "The method can only render one shield if an instruction has multiple shields",
+        message = "The method may or may not render multiple shields for a given instruction",
         replaceWith = ReplaceWith("renderManeuver(maneuver, roadShields)")
     )
     @JvmOverloads
     fun render(maneuver: SubManeuver?, roadShield: RoadShield? = null) {
         val roadShields = ifNonNull(roadShield) {
-            listOf(it)
+            setOf(it)
         }
         renderManeuver(maneuver, roadShields)
     }
@@ -76,7 +76,7 @@ class MapboxSubManeuver : AppCompatTextView {
      * Invoke the method to render sub maneuver instructions
      * @param maneuver SubManeuver
      */
-    fun renderManeuver(maneuver: SubManeuver?, roadShields: List<RoadShield>?) {
+    fun renderManeuver(maneuver: SubManeuver?, roadShields: Set<RoadShield>?) {
         val exitView = MapboxExitText(context)
         exitView.setExitStyle(exitBackground, leftDrawable, rightDrawable)
         val instruction = ManeuverInstructionGenerator.generateSub(